Move analysis tab rendering to templates/package_analysis.dart
